Update body_macro_chen.bas
Name of the file:body_macro_chen.bas
Name of my repo: TuesdayEve
Commit 1: Clarify the macro title, it shows the reader what the macro is about. It also helps to separate this particular from others.
Commit 2: Add requirements of this macro at the beginning of the code, in order to check if all the requirements are met.
Commit 3: Change LeftMargin from (0.64) to (0.63), in order to meet the margin requirement.
Commit 4: Delete repeated Macro, since there are two separated macros doing the same job in the original file
My Code checks obvious errors in the master's code, and  it also improves the macro by adding things like title and requirements.